// src/types/dto/TitleDto.ts
export interface TitleDto {
  id: string;
  title: string;
  image: string;
  siteId: string;
  displayTitle: boolean;
  titleFont: string;
  titlePosition: string;
  displayLogo: boolean;
  logoPosition: string;
  logoUrl: string;
  titleSize: number;
  titleColor: string;
  titleWeight: string;
  generatedUrls: string;
}